\name{asteroid_dens}
\docType{data}
\alias{asteroid_dens}

\title{Densities of asteroids}
\description{
This data set gives measured values for the physical density (in grams/cm^3) of 26 asteroids in the main Asteroid Belt of our Solar System, along with their measurement errors.   This dataset is useful for analysis involving heteroscedastic measurement errors in 1-dimension, such as weighted density estimation and regression.  Further description and references for the dataset are given in  Appendix C.1 of Feigelson & Babu (2012).   
}
\usage{asteroid_dens}
\format{A table containing 26 rows and 3 columns with header row.}
\source{Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press
}
\keyword{datasets}
