\name{COMBO17_lowz}
\docType{data}
\alias{COMBO17_lowz}

\title{Galaxy color-magnitude diagram}
\description{
This data set gives two photometric quantities for 572 low redshift (z < 0.3) galaxies from the COMBO-17 survey (Classifying Objects by Medium-Band Observations in 17 Filters).  The quantities are: absolute magnitude in the blue band, M(B); and ultraviolet-to-blue color index, M(280)-M(B).  This dataset is useful for 2-dimensional density estimation, clustering, and mixture models. Further description is given in  Appendix C.10 of Feigelson & Babu (2012).   The full COMBO-17 dataset can be accessed at \url{http://vizier.u-strasbg.fr/viz-bin/VizieR?-source=II\%2F253A}.  
}
\usage{COMBO17_lowz}
\format{A table containing 572 rows and 2 columns with header row.}
\source{Wolf et al. (2003), Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Wolf, C., Meisenheimer, K., Rix, H.-W., Borch, A., Dye, S. and Kleinheinrich, M. (2003) The COMBO-17 survey: Evolution of the galaxy luminosity function from 25 000 galaxies with 0.2 < z < 1.2, \emph{Astronomy and Astrophysics}, 401, 73-98
}
\keyword{datasets}
