\name{ell_gal_profile}
\docType{data}
\alias{ell_gal_profile}

\title{Elliptical galaxy radial profiles}
\description{
This dataset gives the radial surface brightness profiles for three elliptical galaxies in the nearby Virgo galaxy cluster (Kormendy et al. 2009).  They are useful for exercising nonlinear regression in 2-dimensions; the Sersic (1968) profile function is commonly used.  The samples are described in Appendix C.11 of Feigelson & Babu (2012) with regression illustrated in Chapter 7. 
}
\usage{ell_gal_profile}
\format{A table containing 150 rows and 3 columns with header row. It includes 58 entries for NGC 4472, 52 rows for NGC 4406, and 40 rows for NGC 4551.}
\source{Feigelson and Babu (2012), Kormendy et al. (2009), Sersic (1968)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Kormendy, J., Fisher, D. B., Cornell, M. E. & Bender, R. (2009) Structure and formation of elliptical and spheroidal galaxies, \emph{Astrophysical Journal Supplements}, 182, 216-309 \url{http://adsabs.harvard.edu/abs/2009ApJS..182..216K}

Sersic, J. L. (1968) \emph{Atlas de Galaxias Australes}, Obs. Astronomico de Cordoba \url{http://adsabs.harvard.edu/abs/1968adga.book.....S}
}
\keyword{datasets}
