\name{exoplanet_RV}
\docType{data}
\alias{exoplanet_RV}

\title{Exoplanet radial velocities}
\description{
This dataset gives radial velocity time series for three stars hosting massive planets. The technique of detecting extrasolar planets by periodic Doppler variations in the stellar radial velocity is summarized at \url{http://en.wikipedia.org/wiki/Radial_velocity}.  Data are provided here for three stars: HD 88133 (17 measurements), HD 37124 (52 measurements), and HD 3651 (138 measurements). The four columns give the star identifier, observation times (in modified Julian days), heliocentric radial velocities in meters/sec, and standard deviation of the radial velocities.  References and discussion can be found at \url{http://astrostatistics.psu.edu/datasets/exoplanet_Doppler.html}. This dataset is useful for statistical methods treating detection of non-sinusoical periodicities in sparse irregularly-spaced time series with heteroscedastic measurement errors.   
}
\usage{exoplanet_RV}
\format{A table containing 27 rows and 4 columns}
\keyword{datasets}
