\name{GlobClus_mag}
\docType{data}
\alias{GlobClus_mag}

\title{Globular cluster magnitudes}
\description{
This data set gives near-infrared K-band magnitudes for 81 globular clusters in the Milky Way Galaxy (MWG) and 360 globular clusters in the nearby Andromeda Galaxy (M31).  The data are obtained from Nantais et al. (2006) and the samples are more fully described in Appendix C.3 of Feigelson & Babu (2012).  This dataset is useful for univariate two-sample tests, parametric estimation, and treatment of truncation. 
}
\usage{GlobClus_mag}
\format{A table containing 441 rows and 3 columns with header row. The first column is a subsample indicator.}
\source{Nantais et al. (2006), Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Nantais, J. B., Huchra, J. P., Barmby, P., Olsen, K. A. & Jarrett, T. H. (2006) Nearby spiral globular cluster systems. I. Luminosity functions, \emph{Astronomical Journal}, 131, 1416-1425 \url{http://adsabs.harvard.edu/abs/2006AJ....131.1416N}
}
\keyword{datasets}
