\name{GlobClus_prop}
\docType{data}
\alias{GlobClus_prop}

\title{Galactic globular cluster properties}
\description{
This dataset provides astronomical and astrophysical properties of 147 globular star clusters in the Milky Way Galaxy obtained from the catalog of Webbink (1985). Properties include Galactic location, integrated stellar luminosity, metallicity, ellipticity, central surface brightness, color, and seven measures of dynamical state (core and tidal radius, concentration, central star density and relaxation time, central velocity dispersion and escape velocity).  The sample is described in Appendix C.7 of Feigelson & Babu (2012) with statistical analysis in Chapter 8. It is useful to exercise techniques of multivariate analysis and regression.  
}
\usage{GlobClus_prop}
\format{A table containing 147 rows and 20 columns with header row. Several columns have `NA` entries.}
\source{Feigelson and Babu (2012), Webbink (1985)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Webbink, R. F., Structure parameters of galactic globular clusters, in \emph{Dynamics of Star Clusters}, Dordrecht NL:Reidel, 541-577 \url{http://adsabs.harvard.edu/abs/1985IAUS..113..541W}
}
\keyword{datasets}
