\name{HIP}
\docType{data}
\alias{HIP}

\title{Hipparchos stars}
\description{
This dataset is a subset of the large catalog of stellar positions and parallaxes (inverse of distances) obtained by the European Space Agency's Hipparchos satellite (Perryman et al. 1997).  The 2719 stars are chosen from the Hipparchos Input Catalogue (HIP) with the criterion that the parallax lies between 20 and 25 milliarcsecond (corresponding to distance between 40 and 50 parsecs). This subset includes most of the nearby Hyades star cluster.  The dataset also includes celestial location, proper motion in right ascension and declination, visual magnitude, and \emph{B-V} color index.  The sample is described in Appendix C.6 of Feigelson & Babu (2012).  This dataset is useful for 2- and 3-dimensional clustering, mixture models, regression, and spatial point processes. 
}
\usage{HIP}
\format{A table containing 2719 rows and 9 columns with header row. The last column has a few `NA` entries.}
\source{Feigelson and Babu (2012), Perryman et al. (1997)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Perryman, M. A. C., Lindegren, L., Kovalevsky, J., Hoeg, E., Bastian, U., Bernacca, P,. L., Creze, M., Donati, F., Grenon, M., Grewing, M., van Leeuwen, F., van der Marel, H., Mignard, F., Murray, C. A., Le Poole, R. S., Schrijver, H., Turon, C., Arenou, F., Froeschle, M., and Petersen, C. S. (1997) The HIPPARCOS Catalogue, \emph{Astronomy and Astrophysics}, 323, L49-L52. 
}
\keyword{datasets}
