\name{LMC_dist}
\docType{data}
\alias{LMC_dist}

\title{Distance to the Large Magellanic Cloud}
\description{
This dataset is a collection of 25 measurements of the distance to the Large Magellanic Cloud (LMC), the largest galaxy now orbiting the Milky Way Galaxy (Clementini 2003).  The measurements are made by different research groups using different methods: Cepheid and RR Lyrae luminosities, red clump giants, Supernova 1987a expansion, and so forth.  Each distance modulus (DM) measurement is accompanied by its standard deviation.  The dataset is discussed at \url{http://astrostatistics.psu.edu/datasets/LMC_distance.html}.  This dataset is useful for statistical methods treating heteroscedastic measurement errors.  
}
\usage{LMC_dist}
\format{A table containing 25 rows and 3 columns}
\source{Clementini et al. (2003)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Clementini, G., Gratton, R., Bragaglia, A., Carretta, E., Di Fabrizio, L. and Maio, M.  (2003) Distance to the Large Magellanic Cloud: The RR Lyrae Stars, \emph{Astronomical Journal}, 125, 1309-1329 

}
\keyword{datasets}
