\name{plan_neb_LF}
\docType{data}
\alias{plan_neb_LF}

\title{Planetary nebula luminosity function}
\description{
This dataset contains the visual magnitudes of 531 planetary nebulae in five nearby galaxies: M 31 (Andromeda), M 81, NGC 3379, NGC 4494 and NGC 4382 (Ciardullo et al. 2002 and references therein). If the distribution of planetary nebula luminosities (its `luminosity function') is universal, then the offsets between the distributions can be used to estimate galaxy distances.  However, the samples are truncated at different levels for each galaxy. The dataset is discussed at \url{http://astrostatistics.psu.edu/datasets/plan_neb.html} (with references) and can be used to exercise statistical methods of density estimation and regression in the presence of truncation. 
}
\usage{plan_neb_LF}
\format{A table containing 531 rows and 2 columns with header row}
\source{Ciardullo et al. (2002)}
\references{
Ciardullo, R., Feldmeier, J. J., Jacoby, G. H., Kuzio de Naray, R., Laychak, M. B. and Durrell, P. R. (2002) Planetary nebulae as standard candles. XII. Connecting the Population I and Population II distance scales, \emph{Astrophysical Journal}, 577, 31-50
}
\keyword{datasets}
