\name{protostellar_disks}
\docType{data}
\alias{protostellar_disks}

\title{Protostellar disks}
\description{
This data set gives rxc contingency table of the observed population of 5 evolutionary stages of  pre-main sequence stars in four nearby star forming clouds.  Clouds are listed in order of increasing age, and stellar evolutionary stages are based on the properties of their infrared-emitting circumstellar disks.  Sample sizes in a cell range from 0 to 179 stars. This dataset is useful for analysis involving \emph{r x c} contingency tables with small samples.  Further description and references for the dataset are given in  Appendix C.2 of Feigelson & Babu (2012).   
}
\usage{protostellar_disks}
\format{A table containing 4 rows and 5 columns with header row and column.}
\source{Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

}
\keyword{datasets}
