\name{protostellar_jets}
\docType{data}
\alias{protostellar_jets}

\title{Protostellar jets}
\description{
This data set gives a 2x2 contingency table of the presence of jets produced by protostars in single and multiple systems.  The sample has 21 total systems (Reipurth et al. 2004).  This dataset is useful for analysis involving 2x2 contingency tables of small samples.  Further description and references for the dataset are given in  Appendix C.2 of Feigelson & Babu (2012).   
}
\usage{protostellar_jets}
\format{A table containing 2 rows and 2 columns with header row.}
\source{Reipurth et al. (2004), Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Reipurth, B., Rodriquez, L. F., Anglada, G. & Bally, J. (2004) Radio continuum jets from protostellar objects, \emph{Astronomical Journal}, 127, 1736-1746 \url{http://adsabs.harvard.edu/abs/2004AJ....127.1736R}
}
\keyword{datasets}
