\name{SDSS_QSO}
\docType{data}
\alias{SDSS_QSO}

\title{Sloan Digital Sky Survey quasars}
\description{
This large dataset is the catalog of 77,429 quasars (QSOs) produced from the 5th Data Release of the Sloan Digital Sky Survey (SDSS; York et al. 2000, Schneider et al 2007).  It provides the SDSS designation (which gives the celestial location), redshift (a measure of distance), magnitudes in five photometric bands (with heteroscedastic measurement errors), measures or indicators of radio and X-ray emission,  and absolute magnitude (a measure of luminosity).  The sample is described in Appendix C.8 of Feigelson & Babu (2012) which defines the variables including the radio and X-ray censoring indicators.  The dataset is useful for multivariate analysis and regression including left-censoring.  
}
\usage{SDSS_QSO}
\format{A gzipped table containing 77429 rows and 15 columns with header row. No `NA' entries are present.}
\source{Feigelson and Babu (2012), Schneider et al. (2007), York et al. (2000)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press

Schneider, D. P., Hall, P. B., Richards, G. T., Strauss, M. A., Vanden Berk, D. E. and 39 others (2007) The Sloan Digital Sky Survey Quasar Catalog. IV. Fifth Data Release, \emph{Astronomical Journal}, 134, 102-117 \url{http://adsabs.harvard.edu/abs/2007AJ....134..102S}

York, D. G., Adelman, J., Anderson, J. E., Anderson, S. F., Annis, J., and 140 others (2000) The Sloan Digital Sky Survey: Technical Summary, \emph{Astronomical Journal}, 120, 1579-1587 \url{http://adsabs.harvard.edu/abs/2000AJ....120.1579Y}
}
\keyword{datasets}
