\name{Shapley_galaxy}
\docType{data}
\alias{Shapley_galaxy}

\title{Shapley Concentration of galaxies redshift survey}
\description{
This dataset gives results from a galaxy redshift survey of the Shapley Concentration (SCl 124), the richest nearby supercluster of galaxies centered around (RA,Dec) = (13h25m, -30d).  It includes ~40 Abell rich clusters as well as regions with low galaxy densities (voids).  See \url{http://en.wikipedia.org/wiki/Shapley_Supercluster}.  The dataset, based on the redshift survey of Drinkwater et al. (2004), has 4215 galaxy redshifts with uncertainties) in addition to celestial locations and visible magnitudes.  The sample is described in Appendix C.5 of Feigelson & Babu (2012).  It is useful for study of 3-dimensional spatial point processes with high-amplitude, anisoptropic spatial clustering.  
}
\usage{Shapley_galaxy}
\format{A table containing 4215 rows and 5 columns with header row. Missing data is denoted by zeros.}
\source{Drinkwater et al. 2004, Feigelson and Babu (2012)}
\references{
Drinkwater, M. J., Parker, Q. A., Proust, D., Slezak, E. & Quintana, H. (2004) The large scale distribution of galaxies in the Shapley Supercluster, Publ Astron. Soc. Australia, 21, 89-96 

Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press
}
\keyword{datasets}
