\name{Sun_spot_num}
\docType{data}
\alias{Sun_spot_num}

\title{Sunspot numbers}
\description{
This dataset provides measurements of the number of sunspots, magnetic active regions, seen each month on the Sun from 1749 until May 2014.  It shows the well-known 11-year solar magnetic cycle as well as other variations.  The dataset is described in Appendix C.13 of Feigelson & Babu (2012) and is obtained from the Solar Physics research group at NASA's Marchall Space Flight Center \url{solarscience.msfc.nasa.gov/SunspotCycle.shtml}.  The dataset has four columns: Year, Month, Monthly average sunspot number (SSN), and its standard deviation from the International Sunspot Number compiled by the Solar Influences Data Analysis Center in Belgium \url{http://sidc.oma.be/}.   Users are encouraged to retrieve updated versions prior to study.  The dataset is useful for the analysis of periodic and aperiodic time series.   
}
\usage{Sun_spot_num}
\format{A table containing 3185 rows and 4 columns.}
\source{Feigelson and Babu (2012)}
\references{
Feigelson, E. D. and Babu, G. J. (2012) \emph{Modern Statistical Methods for Astronomy with R Applications}, Cambridge UK:Cambridge University Press
}
\keyword{datasets}
